+++ modules/objdetect/src/aruco2/aruco2_dictionary.hpp
