+++ modules/objdetect/src/aruco2/aruco2_predefined_dictionaries.hpp
